add typeshare feature
