+++ crates/gem_xrp/Cargo.toml
@@ -8,8 +8,8 @@ license = "Apache-2.0"
 repository = "https://github.com/gemwalletcom/core"
 
 [features]
-default = ["models"]
-models = []
+default = []
+typeshare = []
 rpc = []
 
 [dependencies]
