+++ crates/gem_xrp/src/rpc/mod.rs
@@ -1,5 +1,6 @@
 mod client;
 pub mod mapper;
+pub mod model;
 pub mod provider;
 
 pub use self::client::XRPClient;
