+++ crates/gem_xrp/src/typeshare.rs
@@ -1,0 +1,108 @@
+use serde::{Deserialize, Serialize};
+use typeshare::typeshare;
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[typeshare(swift = "Sendable")]
+#[typeshare(swiftGenericConstraints = "T: Sendable")]
+pub struct XRPResult<T> {
+    pub result: T,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[typeshare(swift = "Sendable")]
+pub struct XRPAccountResult {
+    pub account_data: Option<XRPAccount>,
+    pub ledger_current_index: i32,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[typeshare(swift = "Sendable")]
+pub struct XRPAccount {
+    #[serde(rename = "Balance")]
+    pub balance: String,
+    #[serde(rename = "Sequence")]
+    pub sequence: i32,
+    #[serde(rename = "OwnerCount")]
+    pub owner_count: i32,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[typeshare(swift = "Sendable")]
+pub struct XRPFee {
+    pub drops: XRPDrops,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[typeshare(swift = "Sendable")]
+pub struct XRPDrops {
+    pub minimum_fee: String,
+    pub median_fee: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[typeshare(swift = "Sendable")]
+pub struct XRPTransactionBroadcast {
+    pub accepted: Option<bool>,
+    pub engine_result_message: Option<String>,
+    pub error_exception: Option<String>,
+    pub tx_json: Option<XRPTransaction>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[typeshare(swift = "Sendable")]
+pub struct XRPTransaction {
+    pub hash: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[typeshare(swift = "Sendable")]
+pub struct XRPTransactionStatus {
+    pub status: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[typeshare(swift = "Sendable")]
+pub struct XRPLatestBlock {
+    pub ledger_current_index: i32,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[typeshare(swiftGenericConstraints = "T: Sendable")]
+#[typeshare(swift = "Sendable")]
+pub struct XRPAccountObjects<T> {
+    pub account_objects: T,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "PascalCase")]
+#[typeshare(swift = "Sendable")]
+pub struct XRPAccountAsset {
+    pub low_limit: XRPAssetLine,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[typeshare(swift = "Sendable")]
+pub struct XRPAssetLine {
+    pub currency: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[typeshare(swift = "Sendable")]
+pub struct XRPTokenId {
+    pub issuer: String,
+    pub currency: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[typeshare(swift = "Sendable")]
+pub struct XRPAccountLinesResult {
+    pub lines: Option<Vec<XRPAccountLine>>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[typeshare(swift = "Sendable")]
+pub struct XRPAccountLine {
+    pub account: String,
+    pub balance: String,
+    pub currency: String,
+}
